+++ src/messages.js
@@ -1,0 +1,20 @@
+export default {
+  ru: {
+    required: 'Данное поле обязательно для заполнения',
+    minLen: 'Минимально допустимое количество символов равно %s%',
+    maxLen: 'Максимально допустимое количество символов равно %s%',
+    email: 'Поле e-mail имеет неверный формат',
+    phone: 'Телефонный номер имеет неверный формат',
+    number: 'Введенные данные должны быть числом',
+    equalTo: 'Введенные данные не совпадают'
+  },
+  en: {
+    required: 'This field is required',
+    minLen: 'The minimum number of characters allowed is %s%',
+    maxLen: 'The maximum allowed number of characters is %s%',
+    email: 'The e-mail field has the wrong format',
+    phone: 'The phone number is not in the correct format',
+    number: 'The entered data must be a number',
+    equalTo: 'The entered data does not match'
+  }
+};
